make: delete compiling flag '-Wno-sign-compare' delete the '-Wno-sign-compare' of both CFLAGS and CXXFLAGS in file configure.ac, and work out the compiling error in comparison between signed and unsigned integer expression. I still use some type converion(convert 'int' into 'uint'), because changing the variabe type from its definition need too many code modifications.

+++ tests/encodeinput.cpp
@@ -214,7 +214,7 @@ bool EncodeOutput::init(const char* outputFileName, int width , int height)
 
 bool EncodeOutput::write(void* data, int size)
 {
-    return fwrite(data, 1, size, m_fp) == size;
+    return fwrite(data, 1, size, m_fp) == (size_t)size;
 }
 
 const char* EncodeOutputH264::getMimeType()
